# src/web/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flight Tracker API",
    "version": "1.0.0",
    "description": "Real-time aircraft tracking system with zone monitoring and alert capabilities",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4040",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/aircraft": {
      "get": {
        "tags": ["Aircraft"],
        "summary": "Get aircraft observations",
        "description": "Returns current or historical aircraft observations",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Time range like '1h', '24h', '7d'",
            "required": false,
            "schema": {
              "type": "string",
              "example": "24h"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start time (ISO 8601 or timestamp)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End time (ISO 8601 or timestamp)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "description": "Filter by area/zone ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by aircraft type prefix",
            "required": false,
            "schema": {
              "type": "string",
              "example": "B73"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Aircraft"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/observations": {
      "get": {
        "tags": ["Aircraft"],
        "summary": "Get historical observations",
        "description": "Returns aircraft observations with filtering options",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "ISO 8601 date string for start time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "ISO 8601 date string for end time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "areas",
            "in": "query",
            "description": "Comma-separated area IDs",
            "required": false,
            "schema": {
              "type": "string",
              "example": "1,2,3"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Aircraft type filter",
            "required": false,
            "schema": {
              "type": "string",
              "example": "B737"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum results",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Observation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/areas": {
      "get": {
        "tags": ["Areas"],
        "summary": "List all areas",
        "description": "Returns all defined tracking areas (basic info only)",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Area"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Areas"],
        "summary": "Create area from GeoJSON",
        "description": "Create a new tracking area from GeoJSON FeatureCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "GeoJSON FeatureCollection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Area(s) created successfully"
          },
          "400": {
            "description": "Invalid GeoJSON format"
          }
        }
      }
    },
    "/areas/{id}": {
      "delete": {
        "tags": ["Areas"],
        "summary": "Delete area",
        "description": "Delete a tracking area by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Area deleted successfully"
          },
          "404": {
            "description": "Area not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Area": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Area ID"
          },
          "name": {
            "type": "string",
            "description": "Area name"
          },
          "polygon": {
            "type": "string",
            "description": "JSON string of polygon coordinates"
          }
        }
      },
      "Aircraft": {
        "type": "object",
        "properties": {
          "hex": {
            "type": "string",
            "description": "ICAO24 hex code"
          },
          "type": {
            "type": "string",
            "description": "Aircraft type code"
          },
          "callsign": {
            "type": "string",
            "description": "Flight callsign"
          },
          "altitude": {
            "type": "integer",
            "description": "Altitude in feet"
          },
          "speed": {
            "type": "integer",
            "description": "Ground speed in knots"
          },
          "track": {
            "type": "integer",
            "description": "Heading in degrees"
          },
          "lat": {
            "type": "number",
            "description": "Latitude"
          },
          "lon": {
            "type": "number",
            "description": "Longitude"
          },
          "manufacturer": {
            "type": "string",
            "description": "Aircraft manufacturer"
          },
          "model": {
            "type": "string",
            "description": "Aircraft model"
          },
          "operator": {
            "type": "string",
            "description": "Airline or operator"
          },
          "registration": {
            "type": "string",
            "description": "Registration number"
          }
        }
      },
      "Observation": {
        "type": "object",
        "properties": {
          "icao24": {
            "type": "string",
            "description": "ICAO24 hex code"
          },
          "area_id": {
            "type": "integer",
            "description": "Zone ID"
          },
          "enteredAt": {
            "type": "string",
            "format": "date-time",
            "description": "Entry time"
          },
          "exitedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Exit time"
          },
          "type": {
            "type": "string",
            "description": "Aircraft type"
          },
          "altitude": {
            "type": "integer",
            "description": "Altitude in feet"
          },
          "speed": {
            "type": "integer",
            "description": "Speed in knots"
          },
          "manufacturer": {
            "type": "string",
            "description": "Aircraft manufacturer"
          },
          "model": {
            "type": "string",
            "description": "Aircraft model"
          },
          "operator": {
            "type": "string",
            "description": "Airline operator"
          }
        }
      },
...
    }
  }
}